Move makeTemporaryEnv impl into test core
