+++ packages/acceptance-tests/pkg-tests-core/sources/index.ts
@@ -1,5 +1,7 @@
 /// <reference path="../types/global.d.ts" />
 
+import './utils/makeTemporaryEnv';
+
 import * as exec  from './utils/exec';
 import * as fs    from './utils/fs';
 import * as misc  from './utils/misc';
