+++ packages/acceptance-tests/pkg-tests-core/sources/utils/makeTemporaryEnv.ts
@@ -1,0 +1,64 @@
+import {NodeFS}    from '@yarnpkg/fslib';
+import {delimiter} from 'path';
+
+import {URL}       from 'url';
+
+import * as exec   from './exec';
+import * as fs     from './fs';
+import * as tests  from './tests';
+
+const {generatePkgDriver} = tests;
+const {execFile} = exec;
+const {createTemporaryFolder} = fs;
+
+global.makeTemporaryEnv = generatePkgDriver({
+  getName() {
+    return `yarn`;
+  },
+  async runDriver(
+    path,
+    [command, ...args],
+    {cwd, projectFolder, registryUrl, env, ...config},
+  ) {
+    const rcEnv: Record<string, any> = {};
+    for (const key of Object.keys(config))
+      rcEnv[`YARN_${key.replace(/([A-Z])/g, `_$1`).toUpperCase()}`] = config[key];
+
+    const nativePath = NodeFS.fromPortablePath(path);
+    const tempHomeFolder = NodeFS.fromPortablePath(await createTemporaryFolder());
+
+    const cwdArgs = typeof projectFolder !== `undefined`
+      ? [projectFolder]
+      : [];
+
+    const res = await execFile(process.execPath, [`${__dirname}/../../../../../scripts/run-yarn.js`, ...cwdArgs, command, ...args], {
+      cwd: cwd || path,
+      env: {
+        [`HOME`]: tempHomeFolder,
+        [`USERPROFILE`]: tempHomeFolder,
+        [`PATH`]: `${nativePath}/bin${delimiter}${process.env.PATH}`,
+        [`TEST_ENV`]: `true`,
+        [`YARN_GLOBAL_FOLDER`]: `${nativePath}/.yarn/global`,
+        [`YARN_NPM_REGISTRY_SERVER`]: registryUrl!,
+        [`YARN_UNSAFE_HTTP_WHITELIST`]: new URL(registryUrl!).hostname,
+        // Otherwise the tests would break when C:\tmp is on a different drive than the repo
+        [`YARN_ENABLE_ABSOLUTE_VIRTUALS`]: `true`,
+        // Otherwise the output isn't stable between runs
+        [`YARN_ENABLE_TIMERS`]: `false`,
+        // Otherwise we would more often test the fallback rather than the real logic
+        [`YARN_PNP_FALLBACK_MODE`]: `none`,
+        ...rcEnv,
+        ...env,
+      },
+    });
+
+    if (process.env.JEST_LOG_SPAWNS) {
+      console.log(`===== stdout:`);
+      console.log(res.stdout);
+      console.log(`===== stderr:`);
+      console.log(res.stderr);
+    }
+
+    return res;
+  },
+});
