+++ packages/yarnpkg-builder/package.json
@@ -2,7 +2,7 @@
   "name": "@yarnpkg/builder",
   "version": "2.0.0-rc.6",
   "nextVersion": {
-    "nonce": "7935871209988623"
+    "nonce": "3567628816668243"
   },
   "bin": "./sources/boot-dev.js",
   "dependencies": {
